+++ CHANGES.rst
@@ -29,6 +29,7 @@ astropy.io.ascii
 
 astropy.io.misc
 ^^^^^^^^^^^^^^^
+- Added 'functional_models.py' to asdf/tags/transform. Added tags for Gaussian1D and Gaussian2D models to allow serialization. [#9959]
 
 astropy.io.fits
 ^^^^^^^^^^^^^^^